修改受试者实体存在两个gender字段，以及serialNum数据类型更改为Long与数据库一致解决顺序输入总为1的问题

+++ que-center/src/main/java/com/m3ds/que/center/entity/form/SkipForm.java
@@ -52,7 +52,7 @@ public class SkipForm extends BaseForm<Skip> {
     /**
      * 排序序号
      */
-    private Integer serialNum;
+    private Long serialNum;
 
     /**
      * type1:每个conditionJson的map都要指定选项和题目
